user auth and persistent logged-in state complete!
