+++ Tee-Time-Skeleton/auth.js
@@ -7,18 +7,19 @@ const loginUser = (req, res, user) => {
   };
 };
 
+
+
 const logoutUser = (req, res) => {
   delete req.session.auth;
 };
 
-const restoreUser = async (req, res, next) => {
 
+
+const restoreUser = async (req, res, next) => {
   if (req.session.auth) {
     const { userId } = req.session.auth;
-
     try {
       const user = await db.User.findByPk(userId);
-
       if (user) {
         res.locals.authenticated = true;
         res.locals.user = user;
@@ -28,6 +29,7 @@ const restoreUser = async (req, res, next) => {
       res.locals.authenticated = false;
       next(err);
     }
+
   } else {
     res.locals.authenticated = false;
     next();
