+++ Tee-Time-Skeleton/routes/signup.js
@@ -1,117 +1,73 @@
-const express = require('express');
-const { check, validationResult } = require('express-validator');
-const db = require('../db/models');
-const { csrfProtection, asyncHandler } = require('./utils');
-const router = express.Router();
-const bcrypt = require('bcryptjs');
+// const express = require('express');
+// const { check, validationResult } = require('express-validator');
+// const db = require('../db/models');
+// const { csrfProtection, asyncHandler, userValidators } = require('./utils');
+// const router = express.Router();
+// const bcrypt = require('bcryptjs');
 
-router.get('/signup',
-    csrfProtection,
-    asyncHandler(async (req, res) => {
-        const user = {};
-        const cities = await db.City.findAll({ order: ['name'] })
-        res.render('signup', { title: 'Sign Up', user, cities, csrfToken: req.csrfToken(), });
-    })
-);
-
-
-const userValidators = [
-  check('firstName')
-    .exists({ checkFalsy: true })
-    .withMessage('Please provide a value for First Name')
-    .isLength({ max: 50 })
-    .withMessage('First Name must not be more than 50 characters long'),
-  check('lastName')
-    .exists({ checkFalsy: true })
-    .withMessage('Please provide a value for Last Name')
-    .isLength({ max: 50 })
-    .withMessage('Last Name must not be more than 50 characters long'),
-  check('email')
-    .exists({ checkFalsy: true })
-    .withMessage('Please provide a value for Email Address')
-    .isLength({ max: 255 })
-    .withMessage('Email Address must not be more than 255 characters long')
-    .isEmail()
-    .withMessage('Email Address is not a valid email')
-    .custom((value) => {
-      return db.User.findOne({ where: { email: value } })
-        .then((user) => {
-          if (user) {
-            return Promise.reject('The provided Email Address is already in use by another account');
-          }
-        });
-    }),
-  check('password')
-    .exists({ checkFalsy: true })
-    .withMessage('Please provide a value for Password')
-    .isLength({ max: 50 })
-    .withMessage('Password must not be more than 50 characters long')
-    .matches(/^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#$%^&*])/, 'g')
-    .withMessage('Password must contain at least 1 lowercase letter, uppercase letter, number, and special character (i.e. "!@#$%^&*")'),
-  check('confirmPassword')
-    .exists({ checkFalsy: true })
-    .withMessage('Please provide a value for Confirm Password')
-    .isLength({ max: 50 })
-    .withMessage('Confirm Password must not be more than 50 characters long')
-    .custom((value, { req }) => {
-      if (value !== req.body.password) {
-        throw new Error('Confirm Password does not match Password');
-      }
-      return true;
-    }),
-];
+// router.get('/signup',
+//     csrfProtection,
+//     asyncHandler(async (req, res) => {
+//         const user = {};
+//         const cities = await db.City.findAll({ order: ['name'] })
+//         res.render('signup', { title: 'Sign Up', user, cities, csrfToken: req.csrfToken(), });
+//     })
+// );
 
 
-router.post('/signup', 
-  userValidators,
-  csrfProtection,
-  asyncHandler(async (req, res) => {
 
-  const {
-    firstName,
-    lastName,
-    age,
-    bio,
-    cityId,
-    handicap,
-    username,
-    email,
-    password,
-  } = req.body;
 
-  const user = await db.User.build({ 
-    firstName,
-    lastName,
-    age,
-    bio,
-    cityId,
-    handicap,
-    username,
-    email,
-    password 
-  });
 
-  const validatorErrors = validationResult(req);
+// router.post('/signup', 
+//   userValidators,
+//   csrfProtection,
+//   asyncHandler(async (req, res) => {
 
-  if (validatorErrors.isEmpty()) {
-    const hashedPassword = await bcrypt.hash(password, 10);
-    user.hashedPassword = hashedPassword;
-    await user.save();
-    // loginUser(req, res, user);
-    res.redirect(`/users/${user.id}`);
+//   const {
+//     firstName,
+//     lastName,
+//     age,
+//     bio,
+//     cityId,
+//     handicap,
+//     username,
+//     email,
+//     password,
+//   } = req.body;
+
+//   const user = await db.User.build({ 
+//     firstName,
+//     lastName,
+//     age,
+//     bio,
+//     cityId,
+//     handicap,
+//     username,
+//     email,
+//     password 
+//   });
+
+//   const validatorErrors = validationResult(req);
+
+//   if (validatorErrors.isEmpty()) {
+//     const hashedPassword = await bcrypt.hash(password, 10);
+//     user.hashedPassword = hashedPassword;
+//     await user.save();
+//     // loginUser(req, res, user);
+//     res.redirect(`/users/${user.id}`);
     
-  } else {
-    const cities = await db.City.findAll({ order: ['name'] })
-    const errors = validatorErrors.array().map((error) => error.msg);
-    res.render('signup', {
-      title: 'Error',
-      user,
-      errors,
-      cities,
-      csrfToken: req.csrfToken(),
-    });
-  }
-}));
+//   } else {
+//     const cities = await db.City.findAll({ order: ['name'] })
+//     const errors = validatorErrors.array().map((error) => error.msg);
+//     res.render('signup', {
+//       title: 'Error',
+//       user,
+//       errors,
+//       cities,
+//       csrfToken: req.csrfToken(),
+//     });
+//   }
+// }));
 
 
-module.exports = router;
+// module.exports = router;
