+++ Tee-Time-Skeleton/views/user-login.pug
@@ -3,8 +3,13 @@ extends layout.pug
 include utils.pug
 
 block content
+
+  if errors
+    each error in errors
+      .error= error
+      
   +validationErrorSummary(errors)
-  form(action='/user/login' method='post')
+  form(action='/login' method='post')
     input(type='hidden' name='_csrf' value=csrfToken)
     +field('Email Address', 'email', email)
     +field('Password', 'password', null, 'password')
